refactor: extract tool search code controller
