+++ src/agents/tool-search-code-mode-child.ts
@@ -1,0 +1,248 @@
+export const TOOL_SEARCH_CODE_MODE_CHILD_SOURCE = String.raw`
+import vm from "node:vm";
+
+let activeController;
+
+function send(message) {
+  if (typeof process.send === "function" && process.connected) {
+    process.send(message);
+  }
+}
+
+function sendAndFlush(message) {
+  return new Promise((resolve) => {
+    if (typeof process.send !== "function" || !process.connected) {
+      resolve();
+      return;
+    }
+    try {
+      process.send(message, () => resolve());
+    } catch {
+      resolve();
+    }
+  });
+}
+
+function toJsonSafe(value) {
+  if (value === undefined) {
+    return null;
+  }
+  try {
+    return JSON.parse(JSON.stringify(value));
+  } catch {
+    if (value instanceof Error) {
+      return value.message;
+    }
+    if (value === null) {
+      return null;
+    }
+    switch (typeof value) {
+      case "string":
+        return value;
+      case "number":
+      case "boolean":
+      case "bigint":
+      case "symbol":
+      case "function":
+        return String(value);
+      default:
+        return Object.prototype.toString.call(value);
+    }
+  }
+}
+
+function formatLogItem(value) {
+  if (typeof value === "string") {
+    return value;
+  }
+  const safe = toJsonSafe(value);
+  return typeof safe === "string" ? safe : JSON.stringify(safe);
+}
+
+function bridgeResultPayload(message) {
+  if (!message.ok) {
+    return typeof message.error === "string" ? message.error : "tool bridge failed";
+  }
+  const json = JSON.stringify(toJsonSafe(message.value));
+  return typeof json === "string" ? json : "null";
+}
+
+function settleBridge(message) {
+  if (!activeController) {
+    return;
+  }
+  const id = typeof message?.id === "string" ? message.id : "";
+  try {
+    activeController.settleBridge(id, Boolean(message.ok), bridgeResultPayload(message));
+  } catch (error) {
+    send({
+      type: "result",
+      ok: false,
+      error: error instanceof Error ? error.message : String(error),
+    });
+  }
+}
+
+function buildModelScriptSource(code) {
+  return "(async (openclaw, console) => {\n" + code + "\n})(openclaw, console)";
+}
+
+function buildControllerSource() {
+  // The controller returns promise-like bridge handles. The model code can await
+  // them naturally, while the parent process serializes real tool calls.
+  return (
+    '"use strict";\n' +
+    "(() => {\n" +
+    "const pending = new Map();\n" +
+    "const bridgeMessages = [];\n" +
+    "const logs = [];\n" +
+    "let idleWaiters = [];\n" +
+    "let nextBridgeId = 1;\n" +
+    toJsonSafe.toString() +
+    "\n" +
+    formatLogItem.toString() +
+    "\n" +
+    "function notifyBridgeIdle() {\n" +
+    "  if (pending.size !== 0 || bridgeMessages.length !== 0) return;\n" +
+    "  const waiters = idleWaiters;\n" +
+    "  idleWaiters = [];\n" +
+    "  for (const resolve of waiters) resolve();\n" +
+    "}\n" +
+    "function isBridgeIdle() {\n" +
+    "  return pending.size === 0 && bridgeMessages.length === 0;\n" +
+    "}\n" +
+    "function waitForBridgeIdle() {\n" +
+    "  if (isBridgeIdle()) return Promise.resolve();\n" +
+    "  return new Promise((resolve) => idleWaiters.push(resolve));\n" +
+    "}\n" +
+    "function bridge(method, args) {\n" +
+    "  let promise;\n" +
+    "  const start = () => {\n" +
+    "    if (!promise) {\n" +
+    "      const id = String(nextBridgeId++);\n" +
+    "      promise = new Promise((resolve, reject) => {\n" +
+    "        pending.set(id, { resolve, reject });\n" +
+    "        bridgeMessages.push({ id, method, args: toJsonSafe(args) });\n" +
+    "      });\n" +
+    "    }\n" +
+    "    return promise;\n" +
+    "  };\n" +
+    "  return Object.freeze({\n" +
+    "    then: (resolve, reject) => start().then(resolve, reject),\n" +
+    "    catch: (reject) => start().catch(reject),\n" +
+    "    finally: (onFinally) => start().finally(onFinally),\n" +
+    "  });\n" +
+    "}\n" +
+    "const console = Object.freeze({\n" +
+    "  log: (...items) => logs.push(items.map(formatLogItem)),\n" +
+    "  warn: (...items) => logs.push(items.map(formatLogItem)),\n" +
+    "  error: (...items) => logs.push(items.map(formatLogItem)),\n" +
+    "});\n" +
+    "const openclaw = Object.freeze({\n" +
+    "  tools: Object.freeze({\n" +
+    "    search: (query, options) => bridge('search', [query, options]),\n" +
+    "    describe: (id) => bridge('describe', [id]),\n" +
+    "    call: (id, input) => bridge('call', [id, input]),\n" +
+    "  }),\n" +
+    "});\n" +
+    "return Object.freeze({\n" +
+    "  openclaw,\n" +
+    "  console,\n" +
+    "  isBridgeIdle,\n" +
+    "  waitForBridgeIdle,\n" +
+    "  takeLogs: () => logs.splice(0),\n" +
+    "  takeBridgeMessages: () => bridgeMessages.splice(0),\n" +
+    "  settleBridge: (id, ok, payload) => {\n" +
+    "    const waiter = pending.get(String(id));\n" +
+    "    if (!waiter) return;\n" +
+    "    pending.delete(String(id));\n" +
+    "    if (ok) {\n" +
+    "      waiter.resolve(JSON.parse(String(payload)));\n" +
+    "    } else {\n" +
+    "      waiter.reject(new Error(String(payload)));\n" +
+    "    }\n" +
+    "    Promise.resolve().then(notifyBridgeIdle);\n" +
+    "  },\n" +
+    "});\n" +
+    "})()"
+  );
+}
+
+function pumpController(controller) {
+  for (const items of controller.takeLogs()) {
+    send({ type: "log", items });
+  }
+  for (const message of controller.takeBridgeMessages()) {
+    send({ type: "bridge", id: message.id, method: message.method, args: message.args });
+  }
+}
+
+async function runModelCode(code, timeoutMs) {
+  const sandbox = Object.create(null);
+  const context = vm.createContext(sandbox, {
+    name: "tool_search_code",
+    codeGeneration: { strings: false, wasm: false },
+  });
+  const controllerScript = new vm.Script(buildControllerSource(), {
+    filename: "tool_search_code:controller.js",
+  });
+  const controller = controllerScript.runInContext(context, {
+    timeout: Math.max(1, Math.min(Number(timeoutMs) || 1, 2147483647)),
+    breakOnSigint: false,
+  });
+  Object.defineProperties(sandbox, {
+    console: { value: controller.console, enumerable: true },
+    openclaw: { value: controller.openclaw, enumerable: true },
+  });
+  activeController = controller;
+  const pumpTimer = setInterval(() => pumpController(controller), 1);
+  try {
+    const modelScript = new vm.Script(buildModelScriptSource(code), {
+      filename: "tool_search_code:model.js",
+    });
+    const result = await Promise.resolve(
+      modelScript.runInContext(context, {
+        timeout: Math.max(1, Math.min(Number(timeoutMs) || 1, 2147483647)),
+        breakOnSigint: false,
+      }),
+    ).then(
+      (value) => ({ ok: true, value: toJsonSafe(value) }),
+      (error) => ({ ok: false, error: error instanceof Error ? error.message : String(error) }),
+    );
+    do {
+      pumpController(controller);
+      await controller.waitForBridgeIdle();
+      pumpController(controller);
+    } while (!controller.isBridgeIdle());
+    pumpController(controller);
+    await sendAndFlush(
+      result.ok
+        ? { type: "result", ok: true, value: result.value }
+        : { type: "result", ok: false, error: result.error },
+    );
+  } finally {
+    clearInterval(pumpTimer);
+    activeController = undefined;
+  }
+}
+
+process.on("message", (message) => {
+  if (message?.type === "bridge-result") {
+    settleBridge(message);
+    return;
+  }
+  if (message?.type !== "run") {
+    return;
+  }
+  const code = typeof message.code === "string" ? message.code : "";
+  runModelCode(code, message.timeoutMs).catch((error) => {
+    return sendAndFlush({
+      type: "result",
+      ok: false,
+      error: error instanceof Error ? error.message : String(error),
+    });
+  }).finally(() => {
+    setTimeout(() => process.exit(0), 100);
+  });
+});
+`;
